requirements.txt
Django==5.1.4
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.4.2
numpy==1.26.4
scipy==1.13.1
